backend/requirements.txt
# FastAPI & ASGI
fastapi==0.109.2
uvicorn[standard]==0.27.1
gunicorn==21.2.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3

# Validation & Settings
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.9

# Payment
razorpay==1.4.1
stripe==8.2.0

# Email
aiosmtplib==3.0.1

# Middleware
slowapi==0.1.9

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.27.0
aiosqlite==0.20.0

# Utilities
python-dotenv==1.0.1
